Code style tidy-ups for MDC theme interop
- Rename function to MaterialThemeFromAndroidMdcTheme
- Default value for context
- Collapsed TypedArray ext funs to same file
